refactor: move snackbar components from _app into snackbar service provider minor: remove gutters from InvitationList component
