+++ frontend/components/snackbar/MuiSnackbar.test.tsx
@@ -5,8 +5,8 @@
 
 import {render, screen} from '@testing-library/react'
 
-import PageSnackbar from './PageSnackbar'
-import {SnackbarOptions} from './PageSnackbarContext'
+import MuiSnackbar from './MuiSnackbar'
+import {SnackbarOptions} from './MuiSnackbarContext'
 
 //mock snackbar action
 const setSnackbar = jest.fn()
@@ -23,7 +23,7 @@ afterEach(()=>{
 
 it('does NOT render snackbar when open=false',()=>{
   mockOptions.open = false
-  render(<PageSnackbar options={mockOptions} setOptions={setSnackbar} />)
+  render(<MuiSnackbar options={mockOptions} setSnackbar={setSnackbar} />)
   const found = screen.queryByText(mockOptions.message)
   expect(found).toEqual(null)
 })
@@ -32,7 +32,7 @@ it('renders snackbar with mocked ERROR message',()=>{
   mockOptions.severity = 'error'
   mockOptions.message = 'TEST_ERROR_MESSAGE'
   mockOptions.open = true
-  render(<PageSnackbar options={mockOptions} setOptions={setSnackbar} />)
+  render(<MuiSnackbar options={mockOptions} setSnackbar={setSnackbar} />)
   const msg = screen.getByText(mockOptions.message)
   expect(msg).toBeInTheDocument()
 })
@@ -41,7 +41,7 @@ it('renders snackbar with mocked INFO message',()=>{
   mockOptions.severity = 'info'
   mockOptions.message = 'TEST_INFO_MESSAGE'
   mockOptions.open = true
-  render(<PageSnackbar options={mockOptions} setOptions={setSnackbar} />)
+  render(<MuiSnackbar options={mockOptions} setSnackbar={setSnackbar} />)
   const msg = screen.getByText(mockOptions.message)
   expect(msg).toBeInTheDocument()
 })
@@ -50,7 +50,7 @@ it('renders snackbar with mocked SUCCESS message',()=>{
   mockOptions.severity = 'success'
   mockOptions.message = 'TEST_SUCCESS_MESSAGE'
   mockOptions.open = true
-  render(<PageSnackbar options={mockOptions} setOptions={setSnackbar} />)
+  render(<MuiSnackbar options={mockOptions} setSnackbar={setSnackbar} />)
   const msg = screen.getByText(mockOptions.message)
   expect(msg).toBeInTheDocument()
 })
@@ -59,7 +59,7 @@ it('renders snackbar with mocked WARNING message',()=>{
   mockOptions.severity = 'warning'
   mockOptions.message = 'TEST_WARNING_MESSAGE'
   mockOptions.open = true
-  render(<PageSnackbar options={mockOptions} setOptions={setSnackbar} />)
+  render(<MuiSnackbar options={mockOptions} setSnackbar={setSnackbar} />)
   const msg = screen.getByText(mockOptions.message)
   expect(msg).toBeInTheDocument()
 })
